Build as C++17

+++ src/client/ns_turn_msg.c
@@ -567,7 +567,7 @@ int stun_is_challenge_response_str(const uint8_t *buf, size_t len, int *err_code
       const uint8_t *value = stun_attr_get_value(sar);
       if (value) {
         size_t vlen = (size_t)stun_attr_get_len(sar);
-        vlen = min(vlen, STUN_MAX_REALM_SIZE);
+        vlen = min(vlen, (size_t)STUN_MAX_REALM_SIZE);
         memcpy(realm, value, vlen);
         realm[vlen] = 0;
         {
@@ -576,7 +576,7 @@ int stun_is_challenge_response_str(const uint8_t *buf, size_t len, int *err_code
             value = stun_attr_get_value(sar);
             if (value) {
               vlen = (size_t)stun_attr_get_len(sar);
-              vlen = min(vlen, STUN_MAX_SERVER_NAME_SIZE);
+              vlen = min(vlen, (size_t)STUN_MAX_SERVER_NAME_SIZE);
               if (vlen > 0) {
                 if (server_name) {
                   memcpy(server_name, value, vlen);
@@ -592,7 +592,7 @@ int stun_is_challenge_response_str(const uint8_t *buf, size_t len, int *err_code
           value = stun_attr_get_value(sar);
           if (value) {
             vlen = (size_t)stun_attr_get_len(sar);
-            vlen = min(vlen, STUN_MAX_NONCE_SIZE);
+            vlen = min(vlen, (size_t)STUN_MAX_NONCE_SIZE);
             memcpy(nonce, value, vlen);
             nonce[vlen] = 0;
             if (oauth) {
